+++ app/Http/Controllers/ProductController.php
@@ -7,8 +7,8 @@ use App\Models\Product;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Auth;
 use App\Policies\ProductPolicy;
-use Auth;
 use Illuminate\Support\Facades\Storage;
 use App\Models\File;
 use App\Models\Review;
@@ -26,7 +26,6 @@ class ProductController extends Controller
             ->get();
         //error_log("Produtos encontrados: " . $products);
 
-
         return view('pages.products', compact('products'));
     }
 
@@ -37,11 +36,17 @@ class ProductController extends Controller
      */
     public function adminIndex(Request $request)
     {
-        $products = DB::table('product')->where('productname', 'iLIKE', '%' . $request->term . '%')
-            ->get();
-        //error_log("Produtos encontrados: " . $products);
+        $pol = new ProductPolicy();
 
-        return view('admin.products', compact('products'));
+        if ($pol->adminCheck()) {
+            $products = DB::table('product')->where('productname', 'iLIKE', '%' . $request->term . '%')
+                ->get();
+            //error_log("Produtos encontrados: " . $products);
+
+            return view('admin.products', compact('products'));
+        }
+
+        abort(404);
     }
 
     /**
@@ -51,39 +56,51 @@ class ProductController extends Controller
      */
     public function create(Request $request)
     {
-        $product = new Product();
-        $id = auth()->user()->id;
-        if ($id == null)
-            abort(404);
+        $pol = new ProductPolicy();
 
-        $product->seller = $id;
-        $product->productname = $request->input('productname');
-        $product->description = $request->input('description');
-        if (isset($_POST['active']))
-            $product->active = true;
-        else
-            $product->active = false;
+        if ($pol->sellerCheck()) {
+            $product = new Product();
+            $id = auth()->user()->id;
+            if ($id == null)
+                abort(404);
 
-        $path = $request->file('img')->store('public/images');
-        $product->img = $path;
+            $product->seller = $id;
+            $product->productname = $request->input('productname');
+            $product->description = $request->input('description');
+            if (isset($_POST['active']))
+                $product->active = true;
+            else
+                $product->active = false;
 
-        $product->price = $request->input('price');
-        $product->priceperday = $request->input('priceperday');
-        $product->save();
 
-        $category = new Category();
-        $category->product_id = $product->id;
-        $category->name = $request->input('name');
-        $category->save();
 
-        $seller = User::find($product->seller);
+            $file = $request->file('img');
+            $extension = $file->getClientOriginalExtension();
+            $filename = time() . '.' . $extension;
+            $file->move('uploads/productImages', $filename);
+            $product->img = $filename;
 
-        $reviews = DB::table('review')->where('to_user', '=', $seller->id)
-            ->limit(3)
-            ->get();
-        error_log($reviews);
 
-        return view('products.product', ["product" => $product, "seller" => $seller, "reviews" => $reviews]);
+            $product->price = $request->input('price');
+            $product->priceperday = $request->input('priceperday');
+            $product->save();
+
+            $category = new Category();
+            $category->product_id = $product->id;
+            $category->name = $request->input('name');
+            $category->save();
+
+            $seller = User::find($product->seller);
+
+            $reviews = DB::table('review')->where('to_user', '=', $seller->id)
+                ->limit(3)
+                ->get();
+            error_log($reviews);
+
+            return view('products.product', ["product" => $product, "seller" => $seller, "reviews" => $reviews]);
+        }
+
+        abort(404);
     }
 
 
@@ -115,17 +132,35 @@ class ProductController extends Controller
      */
     public function adminShow($id)
     {
-        $product = Product::find($id);
-        if ($product == null)
-            abort(404);
-        return view('admin.productDetails', ["product" => $product]);
+        $pol = new ProductPolicy();
+
+        if ($pol->adminCheck()) {
+            $product = Product::find($id);
+            if ($product == null)
+                abort(404);
+
+            $seller = User::find($product->seller);
+            $reviews = DB::table('review')->where('to_user', '=', $seller->id)
+                ->limit(3)
+                ->get();
+            error_log($reviews);
+            return view('admin.productDetails', ["product" => $product, "seller" => $seller, "reviews" => $reviews]);
+        }
+
+        abort(404);
     }
 
     public function showNewForm()
     {
-        $categories = DB::table('categorynames')
-            ->get();
-        return view('products.new', ["categories" => $categories]);
+        $pol = new ProductPolicy();
+
+        if ($pol->sellerCheck()) {
+            $categories = DB::table('categorynames')
+                ->get();
+            return view('products.new', ["categories" => $categories]);
+        }
+
+        abort(404);
     }
 
     /**
@@ -137,24 +172,35 @@ class ProductController extends Controller
      */
     public function myProducts()
     {
-        $id = auth()->user()->id;
-        if ($id == null)
-            abort(404);
-        $product = DB::table('product')->where('seller', '=', $id)
-            ->get();
+        $pol = new ProductPolicy();
 
-        return view('pages.productManager', ['productManager' => $product]);
+        if ($pol->sellerCheck()) {
+            $id = auth()->user()->id;
+            $product = DB::table('product')->where('seller', '=', $id)
+                ->get();
+
+            return view('pages.productManager', ['productManager' => $product]);
+        }
+
+        abort(404);
     }
 
     public function showUpdateForm($id)
     {
+        $pol = new ProductPolicy();
         $product = Product::find($id);
-        $categories = DB::table('categorynames')
-            ->get();
-        if ($product == null)
-            abort(404);
 
-        return view('products.edit', ["product" => $product], ["categories" => $categories]);
+        if ($pol->updateCheck($product)) {
+
+            $categories = DB::table('categorynames')
+                ->get();
+            if ($product == null)
+                abort(404);
+
+            return view('products.edit', ["product" => $product], ["categories" => $categories]);
+        }
+
+        abort(404);
     }
 
     /**
@@ -166,48 +212,56 @@ class ProductController extends Controller
      */
     public function updatepage(Request $request, $id)
     {
-        $id_aux = auth()->user()->id;
-        $user = User::find($id_aux);
-
+        $pol = new ProductPolicy();
         $product = Product::find($id);
         if ($product == null)
             abort(404);
 
-        if ($request->input('productname') != null)
-            $product->productname = $request->input('productname');
+        if ($pol->updateCheck($product)) {
 
-        if ($request->input('description') != null)
-            $product->description = $request->input('description');
+            if ($request->input('productname') != null)
+                $product->productname = $request->input('productname');
 
-        if (isset($_POST['active']))
-            $product->active = true;
-        else
-            $product->active = false;
+            if ($request->input('description') != null)
+                $product->description = $request->input('description');
 
-        if ($request->input('img') != null)
-            $product->img = $request->input('img');
+            if (isset($_POST['active']))
+                $product->active = true;
+            else
+                $product->active = false;
 
-        if ($request->input('price') != null)
-            $product->price = $request->input('price');
+            if ($request->file('pic') != null) {
+                $file = $request->file('pic');
+                $extension = $file->getClientOriginalExtension();
+                $filename = time() . '.' . $extension;
+                $file->move('uploads/productImages', $filename);
+                $product->img = $filename;
+            }
 
-        if ($request->input('priceperday') != null)
-            $product->priceperday = $request->input('priceperday');
+            if ($request->input('price') != null)
+                $product->price = $request->input('price');
 
-        $product->save();
+            if ($request->input('priceperday') != null)
+                $product->priceperday = $request->input('priceperday');
 
-        if ($request->input('name') != null) {
-            DB::table('category')->where('product_id', '=', $id)->delete();
+            $product->save();
 
-            $category = new Category();
-            $category->product_id = $id;
-            $category->name = $request->input('name');
-            $category->save();
+            if ($request->input('name') != null) {
+                DB::table('category')->where('product_id', '=', $id)->delete();
+
+                $category = new Category();
+                $category->product_id = $id;
+                $category->name = $request->input('name');
+                $category->save();
+            }
+
+            if (auth()->user()->acctype == 'Admin')
+                return redirect('admin/products');
+            else
+                return redirect('products');
         }
 
-        if ($user->acctype == 'Admin')
-            return redirect('admin/products');
-        else
-            return redirect('products');
+        abort(404);
     }
 
     /**
@@ -218,10 +272,16 @@ class ProductController extends Controller
      */
     public function delete($id)
     {
-        $product = Product::find($id);
+        $pol = new ProductPolicy();
 
-        $product->delete();
+        if ($pol->adminCheck()) {
+            $product = Product::find($id);
 
-        return redirect('admin/products');
+            $product->delete();
+
+            return redirect('admin/products');
+        }
+
+        abort(404);
     }
 }
